Move AV utilities to own file.
