+++ opencog/attentionbank/AttentionBank.h
@@ -32,7 +32,6 @@
 
 #include <boost/signals2.hpp>
 
-#include <opencog/util/async_method_caller.h>
 #include <opencog/util/recent_val.h>
 
 #include <opencog/truthvalue/AttentionValue.h>
@@ -44,23 +43,6 @@ namespace opencog
 /** \addtogroup grp_atomspace
  *  @{
  */
-
-/**
- * Handy utilities to get the attention value of an atom.
- */
-AttentionValuePtr get_av(const Handle&);
-AttentionValue::sti_t get_sti(const Handle& h)
-{
-    return get_av(h)->getSTI();
-}
-AttentionValue::lti_t get_lti(const Handle& h)
-{
-    return get_av(h)->getLTI();
-}
-AttentionValue::vlti_t get_vlti(const Handle& h)
-{
-    return get_av(h)->getVLTI();
-}
 
 /* Attention Value changed */
 typedef boost::signals2::signal<void (const Handle&,
